ditto (float -> double changes)
git-svn-id: https://igv.googlecode.com/svn/trunk@1175 17392f64-ead8-4cea-ae29-09b3ab513800

+++ src/org/broad/igv/tools/Preprocessor.java
@@ -579,7 +579,7 @@ public class Preprocessor implements DataConsumer {
         int tileNumber;
         int tileStart;
         int lastFinishedBin = 0;
-        float binWidth;
+        double binWidth;
         int nBins;
         int nonEmptyBins;
         ListAccumulator[][] accumulators;
@@ -592,7 +592,7 @@ public class Preprocessor implements DataConsumer {
             this.tileNumber = tileNumber;
             this.tileStart = tileNumber * tileWidth;
             this.nBins = nBins;
-            this.binWidth = ((float) tileWidth) / nBins;
+            this.binWidth = ((double) tileWidth) / nBins;
             this.accumulators = new ListAccumulator[nTracks][nBins];
         }
